Left board working again; logging moves to see undocumented model
Library's readme was not quite clear on implementation,
fixed a few things by reviewing the codebase and example
implementation in the libary source code.

Taking notes on certain object models, like the move event
emitted by the chess board EventEmitter - is not clear if
this is documented somewhere or if there is a ts model
defined somewhere in the library code.